+++ CHANGELOG.md
@@ -1,6 +1,18 @@
 # Changelog
 
-## [0.0.1] - 2018-8-10
+## [0.0.2] - Aug 8, 2018
+
+### Changed
+
+- memeoftheday.py -> main.py
+- pathHelper.py -> setup.py
+
+### Added
+
+- First time setup if settings.json does not exist
+- Verifies if settings.json is correct and credentials are correct
+
+## [0.0.1] - Aug 8, 2018
 ### Changed
 
 - Changed file structure